fix LongDoubleConversion violation

+++ it/google-cloud-platform/src/test/java/org/apache/beam/it/gcp/bigtable/BigTableIOLT.java
@@ -142,7 +142,7 @@ public class BigTableIOLT extends IOLoadTestBase {
             region,
             readInfo.jobId(),
             getBeamMetricsName(PipelineMetricsType.COUNTER, READ_ELEMENT_METRIC_NAME));
-    assertEquals(configuration.getNumRows(), numRecords, 0.5);
+    assertEquals((double) configuration.getNumRows(), numRecords, 0.5);
 
     // export metrics
     MetricsConfiguration metricsConfig =
